requirements.txt
opencv-python~=4.9.0.80
numpy~=1.26.3
icecream~=2.1.3
pillow~=10.2.0
scipy~=1.12.0
trimesh~=4.1.3
tqdm~=4.66.2
pyhocon~=0.3.60
omegaconf~=2.3.0
packaging~=23.2
einops~=0.7.0
diffusers~=0.26.3
transformers~=4.37.2
imageio~=2.34.0
matplotlib~=3.8.3
datasets~=2.17.0
fire~=0.5.0
gradio~=3.44.0
rembg~=2.0.54
streamlit~=1.31.1
accelerate~=0.27.2
albumentations~=1.3.1
webdataset~=0.2.86
kornia~=0.7.1
pyyaml~=6.0.1
clip
requests~=2.31.0
six~=1.16.0
ipython~=8.21.0
scikit-image~=0.22.0
setuptools~=68.2.2
plotly~=5.18.0
rich~=13.7.0
termcolor~=2.4.0
loguru~=0.7.2
yacs~=0.1.8
flask~=3.0.2
dearpygui~=1.10.1
pymeshlab~=2023.12.post1
plyfile~=1.0.3
psutil~=5.9.8
tensorboardx~=2.6.2.2
torchmetrics~=1.3.1
pandas~=2.2.0
onnxruntime==1.16.0
